Library
This library was made to check to see if strings contained valid input. It is used to make sure that the user is inputting numbers or names correctly. I used it to make sure the input the user is searching for follows the format, and if they edited data, that it too followed the existing format.

This could have simply contained the search methods and sanitization to keep a user an extra degree away from the source data. Given the scope of the original assignment I struggled to come up with reasons for the library. I understand that with large projects they would be very helpful.